# api/tm-catalog.openapi.yaml
openapi: 3.0.3
info:
  title: Thing Model Catalog - OpenAPI 3.0
  version: 0.1.2
  description: "This is the Thing Model Catalog API specification."
  contact: {}  
...
servers:
  - url: ''
security:
  - BearerAuth: []
...
paths:
  /inventory:
    get:
      tags:
        - inventory
      summary: Get the inventory of the catalog
      description: Returns the catalogs inventory
      operationId: getInventory
      parameters:
        - name: 'repo'
          in: query
          description: Source repository name. Optionally constrains the results to only those from given named repository. See '/repos'
          required: false
          schema:
            type: string
          example: 'global'
        - name: 'filter.author'
          in: query
          description: |
            Filters the inventory by one or more authors having exact match.  
            The filter works additive to other filters.
          schema:
            type: string
          example: 'MyCompany,siemens'
        - name: 'filter.manufacturer'
          in: query
          description: |
            Filters the inventory by one or more manufacturers having exact match.  
            The filter works additive to other filters.
          schema:
            type: string
          example: 'BarTech,siemens'
        - name: 'filter.mpn'
          in: query
          description: |
            Filters the inventory by one ore more mpn (manufacturer part number) having exact match.   
            The filter works additive to other filters.
          schema:
            type: string
          example: 'BazLamp,POC1000'
        - name: 'filter.name'
          in: query
          description: |
            Filters the inventory by inventory entry name having a prefix match of full path parts.   
            The filter works additive to other filters.
          schema:
            type: string
          example: 'siemens/siemens/poc1000'
        - name: 'search'
          in: query
          description: |
            Filters the inventory according to whether the content of the inventory entries matches the given search.    
            The search works additive to other filters.
          schema:
            type: string
          example: ''
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryResponse'
              examples:
                inventory:
                  $ref: '#/components/examples/InventoryResponseExample'
        '400':
          description: Invalid filter parameter supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream repository error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /thing-models:
    post:
      tags:
        - thing-models
      summary: Import a Thing Model
      description: Import a Thing Model
      operationId: importThingModel
      parameters:
        - name: force
          in: query
          description: flag to force the import, ignoring any conflicts with existing TMs
          required: false
          schema:
            type: string
        - name: optPath
          in: query
          description: optional path parts to append to the target path (and id) of imported TM, after the mandatory path structure
          required: false
          schema:
            type: string
      requestBody:
        description: |
          Import a Thing Model  
          
          **For the schema of a Thing Model see** [Thing Model JSON schema](https://github.com/w3c/wot-thing-description/blob/main/validation/tm-json-schema-validation.json)
        content:
          application/json:
            schema:
              type: object
            examples:
              payload:
                $ref: '#/components/examples/ImportTMPayloadExample'
        required: true
      responses:
        '201':
          description: Successfully imported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportThingModelResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: Conflict, Thing Model already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    InventoryEntry:
      required:
        - tmName
        - schema:author
        - schema:manufacturer
        - schema:mpn
        - versions
      type: object
      properties:
        tmName:
          type: string
          example: 'siemens/siemens/poc1000'
        schema:author:
          $ref: '#/components/schemas/SchemaAuthor'
        schema:manufacturer:
          $ref: '#/components/schemas/SchemaManufacturer'
        schema:mpn:
          type: string
          example: 'POC1000'
        source:
          $ref: '#/components/schemas/SourceRepository'
        versions:
          type: array
          items:
            $ref: '#/components/schemas/InventoryEntryVersion'
        links:
          $ref: '#/components/schemas/InventoryEntryLinks'
        attachments:
          $ref: "#/components/schemas/AttachmentsList"
...
    InventoryEntryVersion:
      required:
        - tmID
        - externalID
        - description
        - timestamp
        - digest
        - version
      type: object
      properties:
        tmID:
          type: string
          example: 'siemens/siemens/poc1000/v0.0.0-20231201133246-e1594d08a01b.tm.json'
        externalID:
          type: string
          example: 'e7f19370-84a3-11ee-b962-0242ac120002'
        description:
          type: string
          example: 'The Powercenter 1000 is a data transceiver for SENTRON circuit protection devices.'
        digest:
          type: string
          example: 'e1594d08a01b'
        version:
          $ref: '#/components/schemas/ModelVersion'
        timestamp:
          type: string
          example: '20231201133246'
        source:
          $ref: '#/components/schemas/SourceRepository'
        links:
          $ref: '#/components/schemas/InventoryEntryVersionLinks'
        attachments:
          $ref: "#/components/schemas/AttachmentsList"
    InventoryEntryLinks:
      type: object
      required:
        - self
      properties:
        self:
          type: string
          format: uri-reference
          example: './inventory/.tmname/siemens/siemens/poc1000'
    InventoryEntryVersionLinks:
      type: object
      required:
        - self
        - content
      properties:
        self:
          type: string
          format: uri-reference
          example: './inventory/siemens/siemens/poc1000/v0.0.0-20231201133246-e1594d08a01b.tm.json'
        content:
          type: string
          format: uri-reference
          example: './thing-models/siemens/siemens/poc1000/v0.0.0-20231201133246-e1594d08a01b.tm.json'
    InventoryResponse:
      type: object
      required:
        - data
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            $ref: '#/components/schemas/InventoryEntry'
...
    ImportThingModelResponse:
      required:
        - data
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ImportThingModelResult'
...
    AttachmentsList:
      type: array
      items:
        $ref: '#/components/schemas/AttachmentsListEntry'
    AttachmentsListEntry:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        links:
          $ref: '#/components/schemas/AttachmentLinks'
    AttachmentLinks:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          format: uri-reference
          example: './thing-models/siemens/siemens/poc1000/v0.0.0-20231201133246-e1594d08a01b.tm.json/.attachments/README.md'
    SourceRepository:
      type: string
      example: 'global'
      description: |
        The name of the source repository where the inventory entry or version resides. 
        The may be left empty when there is only a single repository served by the backend and thus there is not need for
        disambiguation. See also '/repos'
...
    SchemaAuthor:
      required:
        - schema:name
      type: object
      properties:
        schema:name:
          type: string
          example: 'siemens'
    SchemaManufacturer:
      required:
        - schema:name
      type: object
      properties:
        schema:name:
          type: string
          example: 'siemens'
    Meta:
      type: object
      properties:
        page:
          $ref: '#/components/schemas/MetaPage'
    MetaPage:
      required:
        - elements
      type: object
      properties:
        elements:
          type: integer
    ModelVersion:
      required:
        - model
      type: object
      properties:
        model:
          type: string
          example: '0.0.0'
    ImportThingModelResult:
      required:
        - tmID
      type: object
      properties:
        tmID:
          type: string
          example: 'mycompany/bartech/bazlamp/v0.0.1-20240206122430-1fc13316b7d8.tm.json'
        message:
          type: string
          example: "a TM with the same timestamp but different content exists under ID: mycompany/bartech/bazlamp/v0.0.1-20240206122430-1fc13316b7d8.tm.json"
        code:
          type: string
    ErrorResponse:
      required:
        - title
        - status
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        status:
          type: integer
...
  responses:
    UnauthorizedError:
      description: API key is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  examples:
...
    InventoryResponseExample:
      value:
        meta:
          page:
            elements: 2
        data:
          - 'links':
              'self': './inventory/.tmName/siemens/siemens/poc1000'
            'tmName': 'siemens/siemens/poc1000'
            'schema:author':
              'schema:name': 'siemens'
            'schema:manufacturer':
              'schema:name': 'siemens'
            'schema:mpn': 'POC1000'
            'versions':
              - 'description': 'The Powercenter 1000 is a data transceiver for SENTRON circuit protection devices.'
                'digest': '2cd14601ef09'
                'externalID': '{{ID}}'
                'links':
                  'self': './inventory/siemens/siemens/poc1000/v1.0.0-20240108112117-2cd14601ef09.tm.json'
                  'content': './thing-models/siemens/siemens/poc1000/v1.0.0-20240108112117-2cd14601ef09.tm.json'
                'timestamp': '20240108112117'
                'tmID': 'siemens/siemens/poc1000/v1.0.0-20240108112117-2cd14601ef09.tm.json'
                'version':
                  'model': '1.0.0'
              - 'description': 'The Powercenter 1000 is a data transceiver for SENTRON circuit protection devices.'
                'digest': 'f4184d08a02c'
                'externalID': '{{ID}}'
                'links':
                  'self': './inventory/siemens/siemens/poc1000/v0.0.0-20231217140234-f4184d08a02c.tm.json'
                  'content': './thing-models/siemens/siemens/poc1000/v0.0.0-20231217140234-f4184d08a02c.tm.json'
                'timestamp': '20231217140234'
                'tmID': 'siemens/siemens/poc1000/v0.0.0-20231217140234-f4184d08a02c.tm.json'
                'version':
                  'model': '0.0.0'
              - 'description': ''
                'digest': 'e1594d08a01b'
                'externalID': '{{ID}}'
                'links':
                  'self': './inventory/siemens/siemens/poc1000/v0.0.0-20231201133246-e1594d08a01b.tm.json'
                  'content': './thing-models/siemens/siemens/poc1000/v0.0.0-20231201133246-e1594d08a01b.tm.json'
                'timestamp': '20231201133246'
                'tmID': 'siemens/siemens/poc1000/v0.0.0-20231201133246-e1594d08a01b.tm.json'
                'version':
                  'model': '0.0.0'
          - 'links':
              'self': './inventory/.tmName/MyCompany/BarTech/BazLamp'
            'tmName': 'MyCompany/BarTech/BazLamp'
            'schema:author':
              'schema:name': 'MyCompany'
            'schema:manufacturer':
              'schema:name': 'BarTech'
            'schema:mpn': 'BazLamp'
            'versions':
               - 'description': 'The BarTech BazLamp is a modern smart floor lamp.'
                 'digest': '1fc13316b7d8'
                 'externalID': 'e7f19370-84a3-11ee-b962-0242ac120002'
                 'links':
                   'self': './inventory/MyCompany/BarTech/BazLamp/v0.0.1-20240206122430-1fc13316b7d8.tm.json'
                   'content': './thing-models/MyCompany/BarTech/BazLamp/v0.0.1-20240206122430-1fc13316b7d8.tm.json'
                 'timestamp': '20240206122430'
                 'tmID': 'MyCompany/BarTech/BazLamp/v0.0.1-20240206122430-1fc13316b7d8.tm.json'
                 'version':
                  'model': '0.0.1'
...
    ImportTMPayloadExample:
      value:
        '@context':
           -  'https://www.w3.org/2022/wot/td/v1.1'
           - schema:
               'https://schema.org/'
        '@type': 'tm:ThingModel'
        'id': 'e7f19370-84a3-11ee-b962-0242ac120002'
        'title': 'BazLamp Smart'
        'description': 'The BarTech BazLamp is a modern smart floor lamp.'
        'schema:author':
          'schema:name': 'MyCompany'
        'schema:manufacturer':
          'schema:name': 'BarTech'
        'schema:mpn': 'BazLamp'
        'version':
          'model': 'v0.0.1'
        'properties':
          'status':
            'description': 'current status of the lamp (on|off)'
            'type': 'string'
            'readOnly': true
        'actions':
          'toggle':
            'description': 'Turn the lamp on or off'
        'events':
          'overheating':
            'description': 'Lamp reaches a critical temperature (overheating)'
            'data':
              'type': 'string'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT